latest

+++ src/main/java/com/store/services/ClientService.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.store.domain.Client;
+import com.store.domain.Supplier;
 import com.store.dto.ClientDTO;
 import com.store.enums.StatusEnum;
 import com.store.repositories.ClientRepository;
@@ -15,6 +17,8 @@ public class ClientService{
  
 	@Autowired
 	ClientRepository clientRepository;
+	@Autowired
+	SalesService salesService;
 	
 	public boolean checkifClientexists(ClientDTO clientDTO) {
 	Optional<Client> isClient = clientRepository.findByName1IgnoreCaseAndName2IgnoreCase(clientDTO.getName1(),
@@ -48,6 +52,10 @@ public class ClientService{
 	{
 		return clientRepository.findAllByOrderByIdDesc();	
 	}
+	public List<Client> getAllClientSorted()
+	{
+		return clientRepository.findAll((Sort.by(Sort.Direction.ASC, "name1")));	
+	}
 	public String updateClientById(ClientDTO clientDTO)
 	{
 		    Client updateClient = Client.builder()
@@ -61,9 +69,15 @@ public class ClientService{
 		 	return StatusEnum.CLIENT_UPDATED.value;
 	}
 	public String deleteClientById(long id)
-	{
-		clientRepository.deleteById(id);
-		return StatusEnum.CLIENT_DELETED.value;
+	{  
+		boolean isInSales = salesService.checkIfClientIdExistsInSales(id);
+
+		if (!isInSales) {
+			clientRepository.deleteById(id);
+			return StatusEnum.CLIENT_DELETED.value;
+		} else
+			return StatusEnum.CLIENT_EXIST_IN_SALES.value;
 	}
+	
 	
 }
